+++ test/index.ts
@@ -13,5 +13,6 @@ import "./tag-classification.test";
 import "./lens-presets.test";
 import "./freshness.test";
 import "./gap-finder.test";
+import "./bridge-finder.test";
 
 summary();